Scripts/requirements.txt
azure-search-documents==11.4.0
python-dotenv==1.0.0
pyodbc